Initial navigation components with old home layout
