+++ app/src/main/java/com/sthoray/allright/ui/main/view/MainActivity.kt
@@ -1,22 +1,13 @@
 package com.sthoray.allright.ui.main.view
 
-import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
-import android.view.View
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.GridLayoutManager
+import androidx.navigation.findNavController
+import androidx.navigation.ui.AppBarConfiguration
+import androidx.navigation.ui.setupActionBarWithNavController
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.sthoray.allright.R
-import com.sthoray.allright.data.db.AppDatabase
-import com.sthoray.allright.data.repository.AppRepository
-import com.sthoray.allright.ui.main.adapter.MainAdapter
-import com.sthoray.allright.ui.main.viewmodel.MainViewModel
-import com.sthoray.allright.ui.base.ViewModelProviderFactory
-import com.sthoray.allright.ui.search.view.SearchActivity
-import com.sthoray.allright.utils.Resource
-import kotlinx.android.synthetic.main.activity_main.*
 
 /**
  * The main activity where everything starts.
@@ -26,84 +17,28 @@ import kotlinx.android.synthetic.main.activity_main.*
  */
 class MainActivity : AppCompatActivity() {
 
+    companion object {
 
-    private lateinit var viewModel: MainViewModel
-    private lateinit var mainAdapter: MainAdapter
-    private val TAG = "MainActivity"
-
+        /** The key for a selected categoryId. */
+        const val CATEGORY_ID_KEY = "CATEGORY_ID"
+    }
 
     /**
-     * Set up ViewModel, UI, and observers when the activity is created.
+     * Set up navigation bar.
      *
-     * @param savedInstanceState and data saved if the activity is being re-initialized
+     * @param savedInstanceState If non-null, this activity is being re-constructed
+     * from a previous saved state as given here.
      */
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        setupViewModel()
-        setupUI()
-        setupObservers()
-    }
 
-
-    private fun setupViewModel() {
-        val appRepository = AppRepository(AppDatabase(this))
-        val viewModelProviderFactory =
-            ViewModelProviderFactory(
-                appRepository
-            )
-        viewModel = ViewModelProvider(this, viewModelProviderFactory)
-            .get(MainViewModel::class.java)
-    }
-
-    private fun setupUI() {
-        mainAdapter = MainAdapter()
-        recViewMain.apply {
-            adapter = mainAdapter
-            layoutManager = GridLayoutManager(context, 3)
-        }
-
-        mainAdapter.setOnItemClickListener {
-            val intent = Intent(this, SearchActivity::class.java)
-            intent.putExtra(CATEGORY_ID_KEY, it.id)
-            this.startActivity(intent)
-        }
-    }
-
-    private fun setupObservers() {
-        viewModel.featureCategories.observe(this, Observer { response ->
-            when (response) {
-                is Resource.Success -> {
-                    removeProgressBar()
-                    response.data?.let { featureCategoriesResponse ->
-                        val categories = featureCategoriesResponse.categories.values.toList()
-                        mainAdapter.differ.submitList(categories)
-                    }
-                }
-                is Resource.Error -> {
-                    removeProgressBar()
-                    response.message?.let { message ->
-                        Log.e(TAG, "An error occurred: $message")
-                    }
-                }
-                is Resource.Loading -> {
-                    showProgressBar()
-                }
-            }
-        })
-    }
-
-    private fun showProgressBar() {
-        progBarMainPagination.visibility = View.VISIBLE
-    }
-
-    private fun removeProgressBar() {
-        progBarMainPagination.visibility = View.GONE
-    }
-
-    companion object {
-
-        /** The key for the selected categoryId. */
-        const val CATEGORY_ID_KEY = "CATEGORY_ID"
+        val navView: BottomNavigationView = findViewById(R.id.navigationViewMain)
+        val navController = findNavController(R.id.navHostFragmentMain)
+        val appBarConfiguration = AppBarConfiguration(setOf(
+            R.id.navigation_home, R.id.navigation_browse, R.id.navigation_profile
+        ))
+        setupActionBarWithNavController(navController, appBarConfiguration)
+        navView.setupWithNavController(navController)
     }
 }
